Add userId to the task

+++ backend/src/task/task.controller.ts
@@ -1,4 +1,15 @@
-import { Controller, Get, Post, Put, Delete, Body, Param, UseGuards, Request } from "@nestjs/common"
+import {
+  Controller,
+  Get,
+  Post,
+  Put,
+  Delete,
+  Body,
+  Param,
+  UseGuards,
+  Request,
+  UnauthorizedException,
+} from "@nestjs/common"
 import { TaskService } from "./task.service"
 import { TaskInterface } from "./task.interface"
 import { AuthGuard } from "../auth/auth.guard"
@@ -20,7 +31,13 @@ export class TaskController {
   @UseGuards(AuthGuard)
   @Post("/:id")
   @ApiResponse({ status: 201, description: "Task to POST", type: TaskDTO })
-  async createTask(@Body() task: TaskInterface): Promise<TaskInterface> {
+  async createTask(
+    @Body() task: TaskInterface,
+    @Request() req: RequestWithUser
+  ): Promise<TaskInterface> {
+    if (task.userId !== req.user.userId) {
+      throw new UnauthorizedException()
+    }
     return await this.taskService.createTask(task)
   }
 
@@ -29,8 +46,13 @@ export class TaskController {
   @ApiResponse({ status: 200, description: "Task to PUT", type: TaskDTO })
   async updateTask(
     @Param("id") id: string,
-    @Body() update: Partial<TaskInterface>
+    @Body() update: Partial<TaskInterface>,
+    @Request() req: RequestWithUser
   ): Promise<unknown> {
+    const task = await this.taskService.getTask(id)
+    if (task.userId.toString() !== req.user.userId) {
+      throw new UnauthorizedException()
+    }
     return await this.taskService.updateTask(id, update)
   }
 
@@ -38,8 +60,10 @@ export class TaskController {
   @Delete("/:id")
   @ApiResponse({ status: 200, description: "Task to DELETE", type: TaskDTO })
   async deleteTask(@Param("id") id: string, @Request() req: RequestWithUser): Promise<unknown> {
-    console.log(req.user.userId)
-
+    const task = await this.taskService.getTask(id)
+    if (task.userId.toString() !== req.user.userId) {
+      throw new UnauthorizedException()
+    }
     return await this.taskService.deleteTask(id)
   }
 }
